chore: move compendium browser apps into subfolder
